requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.9
django-cors==0.1
django-filter==23.5
django-restframework==0.0.1
djangorestframework==3.14.0
drf-yasg==1.21.7
idna==3.6
importlib-metadata==7.0.1
inflection==0.5.1
Markdown==3.5.2
packaging==23.2
pytz==2023.4
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.0
zipp==3.17.0
